Split histogram checks from jobs.
